+++ src/sentry_sdk.h
@@ -6,6 +6,7 @@
 #include "sentry/level.h"
 #include "sentry_attachment.h"
 #include "sentry_event.h"
+#include "sentry_logger.h"
 #include "sentry_options.h"
 
 #include <godot_cpp/classes/mutex.hpp>
@@ -32,6 +33,7 @@ private:
 	Ref<RuntimeConfig> runtime_config;
 	Ref<SentryUser> user;
 	Ref<Mutex> user_mutex;
+	Ref<SentryLogger> logger;
 	bool enabled = false;
 	bool configuration_succeeded = false;
 
@@ -39,9 +41,12 @@ private:
 	PackedStringArray _get_global_attachments();
 	void _initialize();
 	void _check_if_configuration_succeeded();
+	void _demo_helper_crash_app();
 
 protected:
 	static void _bind_methods();
+
+	void _notification(int p_what);
 
 public:
 	static SentrySDK *get_singleton() { return singleton; }